requirements.txt
Django==4.2.1
psycopg2-binary==2.9.6
djangorestframework==3.14.0
mysqlclient==2.1.1